+++ e2e/pages/collections/api-documentation/api-documentation.po.js
@@ -1,0 +1,17 @@
+import CollectionsPage from '../collections.po';
+
+const elements = {
+  downloadApiDoc: '#downloadApiDocument',
+};
+
+class ApiDocumentationPage extends CollectionsPage {
+  constructor() {
+    super();
+  }
+
+  get downloadApiDocButton() {
+    return $(elements.downloadApiDoc);
+  }
+}
+
+export default ApiDocumentationPage;