+++ e2e/pages/collections/sed/sed.readonly.spec.js
@@ -1,27 +1,34 @@
-import CollectionsPage from '../collections.po';
+import SedPage from './sed.po';
 import Hooks from '../../../utilities/hooks';
 
 let hooks; let page;
+const path = require('path');
+const fs = require('fs');
+const config = require('../../../config/mainConfig');
 
-describe('the Corrective Action collection page', () => {
+describe('the SED Information for 2015 Edition Products collection page', () => {
   beforeEach(async () => {
-    page = new CollectionsPage();
+    page = new SedPage();
     hooks = new Hooks();
-    hooks.open('#/collections/corrective-action');
+    hooks.open('#/collections/sed');
     await hooks.waitForSpinnerToDisappear();
   });
 
   it('should have body text', () => {
-    expect(page.bodyText.getText()).toContain('This is a list of all health IT products for which a non-conformity has been recorded. A certified product is non-conforming if, at any time, an ONC-Authorized Certification Body (ONC-ACB) or ONC determines that the product does not comply with a requirement of certification. Non-conformities reported as part of surveillance are noted as "Surveillance NCs", while non-conformities identified though an ONC Direct Review are noted as "Direct Review NCs". Not all non-conformities affect a product\'s functionality, and the existence of a non-conformity does not by itself mean that a product is "defective." Developers of certified products are required to notify customers of non-conformities and must take approved corrective actions to address such non-conformities in a timely and effective manner. Detailed information about non-conformities, and associated corrective action plans, can be accessed below by clicking on the product\'s CHPL ID.\nPlease note that by default, only listings that are active or suspended are shown in the search results.');
+    expect(page.bodyText.getText()).toContain('This list includes all 2015 Edition, including Cures Update, health IT products that have been certified with Safety Enhanced Design (SED):');
   });
 
   it('should have table headers in a defined order', () => {
-    const expectedHeaders = ['Edition', 'Developer', 'Product', 'Version', 'CHPL ID', 'ONC-ACB', '# Open Surveillance NCs', '# Closed Surveillance NCs', '# Open Direct Review NCs', '# Closed Direct Review NCs'];
+    const expectedHeaders = ['Developer', 'Product', 'Version', 'CHPL ID', 'Details'];
     const actualHeaders = page.getListingTableHeaders();
     expect(actualHeaders.length).toBe(expectedHeaders.length, 'Found incorrect number of columns');
     actualHeaders.forEach((header, idx) => {
       expect(header.getText()).toBe(expectedHeaders[idx]);
     });
+  });
+
+  it('should have sed details download button', () => {
+    expect(page.sedDetailsDownloadButton.isDisplayed()).toBe(true);
   });
 
   describe('when filtering', () => {
@@ -45,16 +52,6 @@ describe('the Corrective Action collection page', () => {
       });
     });
 
-    describe('using certification edition filter to de select 2015 cures update', () => {
-
-      it('should filter listing results', () => {
-        page.selectFilter('edition', '2015_Cures_Update');
-        page.waitForUpdatedListingResultsCount();
-        countAfter = page.listingTotalCount();
-        expect(countAfter).toBeLessThan(countBefore);
-      });
-    });
-
     describe('using certification status filter to select withdrawn by developer', () => {
 
       it('should filter listing results', () => {
@@ -64,21 +61,11 @@ describe('the Corrective Action collection page', () => {
         expect(countAfter).toBeGreaterThan(countBefore);
       });
     });
-
-    describe('using certification status filter to select open non conformity', () => {
-
-      it('should filter listing results', () => {
-        page.selectFilter('nonconformities', 'open-nonconformity');
-        page.waitForUpdatedListingResultsCount();
-        countAfter = page.listingTotalCount();
-        expect(countAfter).toBeLessThan(countBefore);
-      });
-    });
   });
 
   describe('when searching listing by developer', () => {
-    const DEVELOPER_COL_IDX = 2;
-    const developerName = 'Greenway';
+    const DEVELOPER_COL_IDX = 1;
+    const developerName = 'Allscripts';
     it('should only show listings that match the developer', () => {
       page.searchForListing(developerName);
       page.waitForUpdatedListingResultsCount();
@@ -90,8 +77,8 @@ describe('the Corrective Action collection page', () => {
   });
 
   describe('when searching listing by version', () => {
-    const VERSION_COL_IDX = 4;
-    const versionName = 'v.12';
+    const VERSION_COL_IDX = 3;
+    const versionName = '2017.1.1';
     it('should only show listings that match the version', () => {
       page.searchForListing(versionName);
       page.waitForUpdatedListingResultsCount();
@@ -101,10 +88,10 @@ describe('the Corrective Action collection page', () => {
       }
     });
   });
-
+  
   describe('when searching listing by product', () => {
-    const PRODUCT_COL_IDX = 3;
-    const productName = 'Intergy EHR';
+    const PRODUCT_COL_IDX = 2;
+    const productName = 'Helios';
     it('should only show listings that match the product', () => {
       page.searchForListing(productName);
       page.waitForUpdatedListingResultsCount();
@@ -116,8 +103,8 @@ describe('the Corrective Action collection page', () => {
   });
 
   describe('when searching listing by CHPL ID', () => {
-    const CHPLID_COL_IDX = 5;
-    const chplIdName = '15.04.04.2913.Gree.11.01.1.180919';
+    const CHPLID_COL_IDX = 4;
+    const chplIdName = '15.02.02.1044.A093.01.00.1.190329';
     it('should only show listings that match the product', () => {
       page.searchForListing(chplIdName);
       page.waitForUpdatedListingResultsCount();
@@ -125,6 +112,19 @@ describe('the Corrective Action collection page', () => {
       for (let i = 1; i <= count; i++) {
         expect(page.getColumnText(i, CHPLID_COL_IDX)).toContain(chplIdName);
       }
+    });
+  });
+
+  describe('when clicking on sed details download button', () => {
+
+    it('should download a file', () => {
+      page.sedDetailsDownloadButton.click();
+      let fileName;
+      const sedFileName = 'chpl-sed-all-details';
+      browser.pause(config.timeout); // can't add explicit timeout as file name is dynamic here
+      const files = fs.readdirSync(global.downloadDir);
+      fileName = files.filter((file) => file.match(new RegExp(`${sedFileName}.*.csv`))).toString();
+      expect(fileName).toContain(sedFileName);
     });
   });
 });