+++ src/app/components/filter/saved-filter.component.spec.js
@@ -1,19 +1,23 @@
 (() => {
-  'use strict';
-
   describe('the Saved Filter component', () => {
-
-    var $compile, $log, $q, Mock, ctrl, el, networkService, scope;
+    let $compile;
+    let $log;
+    let $q;
+    let Mock;
+    let ctrl;
+    let el;
+    let networkService;
+    let scope;
 
     beforeEach(() => {
-      angular.mock.module('chpl.mock', 'chpl.components', $provide => {
-        $provide.decorator('networkService', $delegate => {
-          $delegate.getFilters = jasmine.createSpy('getFilters');
-          $delegate.createFilter = jasmine.createSpy('createFilter');
-          $delegate.deleteFilter = jasmine.createSpy('deleteFilter');
-          $delegate.getFilterTypes = jasmine.createSpy('getFilterTypes');
-          return $delegate;
-        });
+      angular.mock.module('chpl.mock', 'chpl.components', ($provide) => {
+        $provide.decorator('networkService', ($delegate) => ({
+          ...$delegate,
+          getFilters: jasmine.createSpy('getFilters'),
+          createFilter: jasmine.createSpy('createFilter'),
+          deleteFilter: jasmine.createSpy('deleteFilter'),
+          getFilterTypes: jasmine.createSpy('getFilterTypes'),
+        }));
       });
 
       inject((_$compile_, _$log_, _$q_, $rootScope, _Mock_, _networkService_) => {
@@ -44,7 +48,7 @@
     afterEach(() => {
       if ($log.debug.logs.length > 0) {
         /* eslint-disable no-console,angular/log */
-        console.log('Debug:\n' + $log.debug.logs.map(o => angular.toJson(o)).join('\n'));
+        console.log(`Debug:\n${$log.debug.logs.map((o) => angular.toJson(o)).join('\n')}`);
         /* eslint-enable no-console,angular/log */
       }
     });
@@ -60,24 +64,33 @@
         expect(ctrl.availableFilters.length).toBe(3);
       });
 
-      it('when saving a filter', () => {
+      it('should give errors when saving a filter with no name', () => {
+        ctrl.saveFilter();
+        expect(scope.createFilterDataObject).not.toHaveBeenCalled();
+        expect(networkService.createFilter).not.toHaveBeenCalled();
+        expect(ctrl.errorMessage).toBe('Filter name is required');
+      });
+
+      it('should succeed when saving a filter', () => {
+        ctrl.filterName = 'a new name';
         ctrl.saveFilter();
         expect(scope.createFilterDataObject).toHaveBeenCalled();
         expect(networkService.createFilter).toHaveBeenCalled();
+        expect(ctrl.errorMessage).toBeUndefined();
       });
 
-      it('when refreshing the list of filters', () => {
+      it('should get filters from the network on refresh', () => {
         ctrl.refreshFilterList();
         expect(networkService.getFilters).toHaveBeenCalled();
       });
 
-      it('when deleteing a filters', () => {
+      it('should use the network service to deleting filters', () => {
         ctrl.deleteFilter();
         expect(networkService.deleteFilter).toHaveBeenCalled();
       });
 
-      it('when selecting a filter', () => {
-        let filter = {filter: {test: 'test'}};
+      it('should apply filters when one is selected when selecting a filter', () => {
+        const filter = { filter: { test: 'test' } };
         ctrl.applyFilter(filter);
         expect(scope.onApplyFilter).toHaveBeenCalledWith(filter.filter);
       });