+++ RELEASE_NOTES.md
@@ -1,5 +1,36 @@
 # Release Notes
 
+## Version 11.14.0
+_2 June 2021_
+
+### Features
+* Update README documentation
+* Add developer column to version activity report
+
+### Bug Fixes
+* Require name when saving filters
+* Show G1/G2 measure changes in Listing history component
+* Filter Listings by Compliance correctly when "Matching All" is selected
+
+### User Interface Updates
+* Display Listing History in updated style
+
+---
+
+## Version 11.13.0
+_17 May 2021_
+
+### Features
+* Display more Listing data in Complaints download file
+* Use new style of component for Meaningful Use User Information Upload
+* Show & allow edit of Service Base URL List value
+* Remove obsolete "Background Jobs" administration page
+
+### Bug Fixes
+* Create link between ICS source/inherited listings correctly
+
+---
+
 ## Version 11.12.0
 _4 May 2021_